requirements.txt
#pip install -r requirements.txt
pykernel
streamlit
langchain
langchain_community
langchain-text-splitters 
pypdf
bs4
python-dotenv
langchain_openai
faiss-cpu
langchain_ollama
streamlit